+++ crates/api-desc/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Patch
 
+- Add `crypto::gcm::tag_length()` function
 - Use `*const u8` instead of `*mut u8` for opaque data
 
 ## 0.1.3